Various Bugfixes

+++ client/src/gfx/shader.c
@@ -32,46 +32,46 @@ shader *sRain;
 shader *sTextMesh;
 shader *sCloud;
 
-extern        size_t src_shader_blockShaderFS_glsl_len;
+extern unsigned  int src_shader_blockShaderFS_glsl_len;
 extern unsigned char src_shader_blockShaderFS_glsl_data[];
 
-extern        size_t src_shader_blockShaderVS_glsl_len;
+extern unsigned  int src_shader_blockShaderVS_glsl_len;
 extern unsigned char src_shader_blockShaderVS_glsl_data[];
 
-extern        size_t src_shader_cloudShaderFS_glsl_len;
+extern unsigned  int src_shader_cloudShaderFS_glsl_len;
 extern unsigned char src_shader_cloudShaderFS_glsl_data[];
 
-extern        size_t src_shader_cloudShaderVS_glsl_len;
+extern unsigned  int src_shader_cloudShaderVS_glsl_len;
 extern unsigned char src_shader_cloudShaderVS_glsl_data[];
 
-extern        size_t src_shader_meshShaderFS_glsl_len;
+extern unsigned  int src_shader_meshShaderFS_glsl_len;
 extern unsigned char src_shader_meshShaderFS_glsl_data[];
 
-extern        size_t src_shader_meshShaderVS_glsl_len;
+extern unsigned  int src_shader_meshShaderVS_glsl_len;
 extern unsigned char src_shader_meshShaderVS_glsl_data[];
 
-extern        size_t src_shader_particleShaderFS_glsl_len;
+extern unsigned  int src_shader_particleShaderFS_glsl_len;
 extern unsigned char src_shader_particleShaderFS_glsl_data[];
 
-extern        size_t src_shader_particleShaderVS_glsl_len;
+extern unsigned  int src_shader_particleShaderVS_glsl_len;
 extern unsigned char src_shader_particleShaderVS_glsl_data[];
 
-extern        size_t src_shader_rainShaderFS_glsl_len;
+extern unsigned  int src_shader_rainShaderFS_glsl_len;
 extern unsigned char src_shader_rainShaderFS_glsl_data[];
 
-extern        size_t src_shader_rainShaderVS_glsl_len;
+extern unsigned  int src_shader_rainShaderVS_glsl_len;
 extern unsigned char src_shader_rainShaderVS_glsl_data[];
 
-extern        size_t src_shader_shadowShaderFS_glsl_len;
+extern unsigned  int src_shader_shadowShaderFS_glsl_len;
 extern unsigned char src_shader_shadowShaderFS_glsl_data[];
 
-extern        size_t src_shader_shadowShaderVS_glsl_len;
+extern unsigned  int src_shader_shadowShaderVS_glsl_len;
 extern unsigned char src_shader_shadowShaderVS_glsl_data[];
 
-extern        size_t src_shader_textShaderFS_glsl_len;
+extern unsigned  int src_shader_textShaderFS_glsl_len;
 extern unsigned char src_shader_textShaderFS_glsl_data[];
 
-extern        size_t src_shader_textShaderVS_glsl_len;
+extern unsigned  int src_shader_textShaderVS_glsl_len;
 extern unsigned char src_shader_textShaderVS_glsl_data[];
 
 void shaderInit(){
