+++ client/src/gfx/texture.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "../../../common/src/common.h"
 
-texture *textureNew               (const u8 *data, size_t dataLen,const char *filename);
-texture *textureNewArray          (const u8 *data, size_t dataLen,const char *filename, int d);
+texture *textureNew               (const u8 *data, unsigned int dataLen,const char *filename);
+texture *textureNewArray          (const u8 *data, unsigned int dataLen,const char *filename, int d);
 void     textureFree              ();
 void     textureBind              (const texture *tex);
 void     textureInit              ();
