+++ client/src/sdl/sfx.h
@@ -20,7 +20,7 @@ struct sfx {
 
 extern sfx sfxList[32];
 
-sfx *sfxNew    (const void *data,size_t dataLen, const char *lName);
+sfx *sfxNew    (const void *data, unsigned int dataLen, const char *lName);
 void sfxFree   (sfx *b);
 void sfxPlay   (sfx *b, float volume);
 void sfxLoop   (sfx *b, float volume);
